tests passing on ios & android
